+++ strider/merge_tracks/__init__.py
@@ -2,4 +2,3 @@ from strider.merge_tracks.rules import rule, action, trigger,\
     source_re, add_tags, change_id, has_tag,\
     SKIP
 from strider.merge_tracks.load_rules import load_rule_file
-# todo document this package